+++ app/(public)/layout.tsx
@@ -1,14 +1,58 @@
+import "./components/styles/globals.css"
+
 import LibraryFooter from "@/components/layout/library-footer";
 import GlobalFooter from "@/components/layout/global-footer";
 import Header from "@/components/layout/header";
+import Editori11y from "@/components/editori11y";
+import Script from "next/script";
+import GoogleAnalytics from "@/components/utils/google-analytics";
 import {ReactNode} from "react";
+import {Icon} from "next/dist/lib/metadata/types/metadata-types";
 
-const Layout = ({children}: { children: ReactNode }) => {
+const appleIcons: Icon[] = [60, 72, 76, 114, 120, 144, 152, 180].map(size => ({
+  url: `https://www-media.stanford.edu/assets/favicon/apple-touch-icon-${size}x${size}.png`,
+  sizes: `${size}x${size}`,
+}));
+
+const icons: Icon[] = [16, 32, 96, 128, 192, 196].map(size => ({
+  url: size === 128 ? `https://www-media.stanford.edu/assets/favicon/favicon-${size}.png` : `https://www-media.stanford.edu/assets/favicon/favicon-${size}x${size}.png`,
+  sizes: `${size}x${size}`
+}));
+
+export const metadata = {
+  title: process.env.NEXT_PUBLIC_SITE_NAME,
+  openGraph: {
+    type: 'website',
+    locale: 'en_IE',
+    url: 'https://library.stanford.edu',
+    siteName: process.env.NEXT_PUBLIC_SITE_NAME,
+  },
+  twitter: {
+    card: 'summary_large_image',
+  },
+  icons: {
+    icon: [{url: '/favicon.ico'}, ...icons],
+    apple: appleIcons
+  }
+}
+
+const RootLayout = ({children}: { children: ReactNode }) => {
 
   return (
+    <html lang="en">
+    <Editori11y/>
+    {process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID &&
+      <>
+        <Script async src="//siteimproveanalytics.com/js/siteanalyze_80352.js"/>
+        <GoogleAnalytics/>
+      </>
+    }
+
+    <body>
 
     <div className="su-grid su-grid-rows-1 su-min-h-screen">
       <div>
+        <a className="su-skiplink" href="#main-content">Skip to main content</a>
         {/* @ts-expect-error Async Server Component */}
         <Header/>
 
@@ -22,6 +66,8 @@ const Layout = ({children}: { children: ReactNode }) => {
         <GlobalFooter/>
       </footer>
     </div>
+    </body>
+    </html>
   )
 }
-export default Layout;
+export default RootLayout;
